+++ .storybook/main.ts
@@ -19,7 +19,6 @@ const config: StorybookConfig = {
   addons: [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
-    "@storybook/addon-onboarding",
     "@storybook/addon-interactions",
     "storybook-dark-mode",
     "@storybook/addon-webpack5-compiler-swc",