+++ .storybook/theme/theme.tsx
@@ -15,12 +15,18 @@ export const ThemeDocsProvider: React.FC<any> = (props) => {
     return () => channel.removeListener(DARK_MODE_EVENT_NAME, setDark);
   }, [channel, setDark]);
 
-  return <DocsContainer {...props} theme={isDark ? dark : light} />;
+  return (
+    // <ThemeProvider theme={isDark ? "dark" : "base"}>
+    <DocsContainer {...props} theme={isDark ? dark : light} />
+    // </ThemeProvider>
+  );
 };
 
 export const ThemeRaruiProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
+  console.log("useDarkMode()", useDarkMode());
+
   return (
     <ThemeProvider theme={useDarkMode() ? "dark" : "base"}>
       {children}