+++ packages/implementations/react/.storybook/preview.tsx
@@ -1,3 +1,3 @@
-import preview from "../../../../.storybook/preview";
+import preview from ".storybook/preview";
 
 export default preview;