+++ packages/implementations/react/src/index.ts
@@ -1,5 +1,8 @@
 // atomics
 export * from "@rarui-react/box";
 export * from "@rarui-react/button";
+export * from "@rarui-react/icon";
 export * from "@rarui-react/link";
+export * from "@rarui-react/skeleton";
+export * from "@rarui-react/text";
 export * from "@rarui-react/title";