+++ packages/styles/src/index.ts
@@ -1,17 +1,23 @@
 // components
 export * from "./components";
 
-// atomics
+// atomic
 export { badge } from "./packages/atomic/badge";
 export { box } from "./packages/atomic/box";
 export { button } from "./packages/atomic/button";
 export { chip } from "./packages/atomic/chip";
+export { icon } from "./packages/atomic/icon";
 export { link } from "./packages/atomic/link";
+export { skeleton } from "./packages/atomic/skeleton";
+export { text } from "./packages/atomic/text";
 export { title } from "./packages/atomic/title";
 export { checkbox } from "./packages/atomic/checkbox";
 
 // types
 export type { BoxSprinkle } from "./packages/atomic/box";
+export type { IconSprinkle } from "./packages/atomic/icon";
+export type { TextSprinkle } from "./packages/atomic/text";
+export type { SkeletonSprinkle } from "./packages/atomic/skeleton";
 export type { TitleSprinkle } from "./packages/atomic/title";
 export type { ButtonVariants } from "./packages/atomic/button";
 export type { CheckboxVariants } from "./packages/atomic/checkbox";