+++ packages/styles/src/packages/atomic/text/index.ts
@@ -1,0 +1,8 @@
+import { textSprinkle, styles } from "./rarui-text.css";
+
+export const text = {
+  classnames: { ...styles },
+  ...textSprinkle,
+};
+
+export type { TextSprinkle } from "./rarui-text.css";