+++ packages/implementations/react/src/atomic/Button/CHANGELOG.md
@@ -6,8 +6,8 @@ The Button component allows us to initiate actions, make state or page changes.
 
 #### 🎉 New features
 
-- Added `children`, `appearance`, `size`, `variant` and `disabled` properties to the Component. ([#5](https://git.rarolabs.com.br/frontend/rarui/pull/5) by [@junior](https://git.rarolabs.com.br/junior))
-- Added stories on Component. ([#5](https://git.rarolabs.com.br/frontend/rarui/pull/5) by [@junior](https://git.rarolabs.com.br/junior))
+- Added `children`, `appearance`, `size`, `variant` and `disabled` properties to the Component. ([#5](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/5) by [@junior](https://git.rarolabs.com.br/junior))
+- Added stories on Component. ([#5](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/5) by [@junior](https://git.rarolabs.com.br/junior))
 
 <!-- #### 🛠 Breaking changes -->